backend/requirements.txt
# --- Web framework ---
fastapi
uvicorn

# --- File upload ---
python-multipart

# --- Machine learning (resume classifier) ---
scikit-learn==1.8.0
joblib

# --- PDF text extraction ---
pdfplumber

# --- LLM integration ---
openai



python-dotenv